Extract `NumberStepperInput` component. Use in multisig threshold input
